code_analyzer_app/requirements.txt
langchain
langchain-openai
langchain-community
tiktoken
python-dotenv
tqdm
pylint # Optional, for static analysis
bandit # Optional, for static analysis